Revert "Merge pull request #1948 from HDR-Development/Super-Fighting-Robot"
This reverts commit ad8996eb2c94129fe3ea60960f34797d8720c438, reversing
changes made to d0c36710a4758d65aa50310d905ac99569ff361f.
