+++ fighters/rockman/src/opff.rs
@@ -36,19 +36,27 @@ use globals::*;
 // }
 
 // Jump cancel dtilt on hit
-// unsafe fn jc_dtilt_hit(boma: &mut BattleObjectModuleAccessor, status_kind: i32, situation_kind: i32, cat1: i32, frame: f32) {
-//     if boma.is_motion(Hash40::new("attack_lw3")) {
-//         if (AttackModule::is_infliction_status(boma, *COLLISION_KIND_MASK_HIT) && !boma.is_in_hitlag()) && frame > 12.0 {
-//             boma.check_jump_cancel(false, false);
-//         }
-//     }
-// }
+unsafe fn jc_dtilt_hit(boma: &mut BattleObjectModuleAccessor, status_kind: i32, situation_kind: i32, cat1: i32, frame: f32) {
+    if boma.is_motion(Hash40::new("attack_lw3")) {
+        if (AttackModule::is_infliction_status(boma, *COLLISION_KIND_MASK_HIT) && !boma.is_in_hitlag()) && frame > 12.0 {
+            boma.check_jump_cancel(false, false);
+        }
+    }
+}
 
 // Mega Man Metal Blad Toss Airdodge Cancel
 unsafe fn blade_toss_ac(boma: &mut BattleObjectModuleAccessor, status_kind: i32, situation_kind: i32, cat1: i32, frame: f32) {
-    if status_kind == *FIGHTER_STATUS_KIND_SPECIAL_S {
+    if status_kind == *FIGHTER_STATUS_KIND_SPECIAL_N {
         if boma.status_frame() > 16 {
             boma.check_airdodge_cancel();
+        }
+    }
+}
+
+unsafe fn side_special_drift(boma: &mut BattleObjectModuleAccessor, status_kind: i32, situation_kind: i32, cat2: i32, stick_y: f32) {
+    if status_kind == *FIGHTER_STATUS_KIND_SPECIAL_S && situation_kind == *SITUATION_KIND_AIR {
+        if KineticModule::get_kinetic_type(boma) != *FIGHTER_KINETIC_TYPE_FALL {
+            KineticModule::change_kinetic(boma, *FIGHTER_KINETIC_TYPE_FALL);
         }
     }
 }
@@ -57,6 +65,7 @@ unsafe fn fastfall_specials(fighter: &mut L2CFighterCommon) {
     if !fighter.is_in_hitlag()
     && !StatusModule::is_changing(fighter.module_accessor)
     && fighter.is_status_one_of(&[
+        *FIGHTER_STATUS_KIND_SPECIAL_S,
         *FIGHTER_STATUS_KIND_SPECIAL_LW,
         *FIGHTER_ROCKMAN_STATUS_KIND_SPECIAL_HI_JUMP,
         *FIGHTER_ROCKMAN_STATUS_KIND_SPECIAL_LW_SHOOT
@@ -86,8 +95,9 @@ unsafe fn fastfall_specials(fighter: &mut L2CFighterCommon) {
 pub unsafe fn moveset(fighter: &mut L2CFighterCommon, boma: &mut BattleObjectModuleAccessor, id: usize, cat: [i32 ; 4], status_kind: i32, situation_kind: i32, motion_kind: u64, stick_x: f32, stick_y: f32, facing: f32, frame: f32) {
     // light_utilt_cancel(boma, id, status_kind, situation_kind, cat[0], frame);
     // utilt_command_input(boma, id, status_kind, situation_kind, frame);
-    // jc_dtilt_hit(boma, status_kind, situation_kind, cat[0], frame);
+    jc_dtilt_hit(boma, status_kind, situation_kind, cat[0], frame);
     blade_toss_ac(boma, status_kind, situation_kind, cat[0], frame);
+    side_special_drift(boma, status_kind, situation_kind, cat[1], stick_y);
     fastfall_specials(fighter);
 }
