+++ fighters/rockman/src/status.rs
@@ -1,77 +1,53 @@
 use super::*;
 use globals::*;
+ 
 
-pub mod helper;
-mod walk;
-mod attack;
-mod attack_s3;
-mod attack_s4;
+// FIGHTER_STATUS_KIND_WALK
 
-mod attack_air;
-mod ladder_attack;
+#[status_script(agent = "rockman", status = FIGHTER_STATUS_KIND_WALK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_PRE)]
+pub unsafe fn pre_walk(fighter: &mut L2CFighterCommon) -> L2CValue {
+    let ground_brake = WorkModule::get_param_float(fighter.module_accessor, hash40("ground_brake"), 0);
 
-mod special_n;
-mod rockbuster;
-mod chargeshot;
+	let mut initial_speed = VarModule::get_float(fighter.battle_object, vars::common::instance::CURR_DASH_SPEED);
 
-mod special_s;
+	if ![*FIGHTER_STATUS_KIND_DASH, *FIGHTER_STATUS_KIND_TURN_DASH].contains(&StatusModule::prev_status_kind(fighter.module_accessor, 0)) {
+		// println!("not after dash/turndash");
+		initial_speed = KineticModule::get_sum_speed_x(fighter.module_accessor, *KINETIC_ENERGY_RESERVE_ATTRIBUTE_ALL) - KineticModule::get_sum_speed_x(fighter.module_accessor, *KINETIC_ENERGY_RESERVE_ATTRIBUTE_GROUND) - KineticModule::get_sum_speed_x(fighter.module_accessor, *KINETIC_ENERGY_RESERVE_ATTRIBUTE_EXTERN);
+	}
 
-mod special_lw;
+	// println!("walk initial speed: {}", initial_speed);
 
-unsafe extern "C" fn rockman_check_special_uniq(fighter: &mut L2CFighterCommon) -> L2CValue {
-    if VarModule::is_flag(fighter.battle_object, vars::rockman::instance::CHARGE_SHOT_RELEASE)
-    && VarModule::is_flag(fighter.battle_object, vars::rockman::instance::CHARGE_SHOT_PLAYED_FX) {
-        fighter.global_table[CMD_CAT1].assign(&L2CValue::I32(*FIGHTER_PAD_CMD_CAT1_FLAG_SPECIAL_N));
-    }
-    false.into()
+	VarModule::set_float(fighter.battle_object, vars::common::instance::CURR_DASH_SPEED, initial_speed);
+
+    original!(fighter)
 }
 
-unsafe extern "C" fn rockman_special_lw_uniq(fighter: &mut L2CFighterCommon) -> L2CValue {
-    (!WorkModule::is_flag(fighter.module_accessor, *FIGHTER_ROCKMAN_INSTANCE_WORK_ID_FLAG_SPECIAL_LW_LEAFSHIELD)).into()
+#[status_script(agent = "rockman", status = FIGHTER_STATUS_KIND_WALK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
+pub unsafe fn walk(fighter: &mut L2CFighterCommon) -> L2CValue {
+    WorkModule::on_flag(fighter.module_accessor, *FIGHTER_STATUS_WALK_FLAG_SLIP);
+    WorkModule::enable_transition_term_group(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_GROUP_CHK_GROUND_SPECIAL);
+    WorkModule::enable_transition_term_group(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_GROUP_CHK_GROUND_ITEM);
+    WorkModule::enable_transition_term_group(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_GROUP_CHK_GROUND_CATCH);
+    WorkModule::enable_transition_term_group(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_GROUP_CHK_GROUND_ATTACK);
+    WorkModule::enable_transition_term_group(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_GROUP_CHK_GROUND_ESCAPE);
+    WorkModule::enable_transition_term_group(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_GROUP_CHK_GROUND_GUARD);
+    WorkModule::enable_transition_term_group(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_GROUP_CHK_GROUND_JUMP);
+    WorkModule::enable_transition_term_group(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_GROUP_CHK_GROUND);
+    WorkModule::unable_transition_term_group_ex(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_TERM_ID_CONT_TURN);
+    WorkModule::unable_transition_term_group_ex(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_TERM_ID_CONT_WALK);
+    WorkModule::unable_transition_term(fighter.module_accessor, *FIGHTER_STATUS_TRANSITION_TERM_ID_SLIP);
+
+    fighter.sub_shift_status_main(L2CValue::Ptr(walk_main as *const () as _))
 }
 
-#[fighter_reset]
-fn agent_reset(fighter: &mut L2CFighterCommon) {
-    unsafe {
-        let fighter_kind = utility::get_kind(&mut *fighter.module_accessor);
-        if fighter_kind != *FIGHTER_KIND_ROCKMAN {
-            return;
-        }
-        fighter.global_table[0x27].assign(&L2CValue::Ptr(rockman_check_special_uniq as *const () as _));
-        fighter.global_table[0x26].assign(&L2CValue::Ptr(rockman_check_special_uniq as *const () as _));
-        fighter.global_table[DASH_CALLBACK].assign(&false.into());
-        fighter.global_table[0x2A].assign(&false.into());
-        fighter.global_table[0x2B].assign(&false.into());
-        fighter.global_table[0x34].assign(&false.into());
-        fighter.global_table[0x35].assign(&false.into());
-        fighter.global_table[0x4E].assign(&false.into());
-        fighter.global_table[USE_SPECIAL_LW_CALLBACK].assign(&L2CValue::Ptr(rockman_special_lw_uniq as *const () as _));
-    }
+unsafe extern "C" fn walk_main(fighter: &mut L2CFighterCommon) -> L2CValue {
+    fighter.status_Walk_Main();
+    0.into()
 }
 
 pub fn install() {
-    install_agent_resets!(
-        agent_reset
+    install_status_scripts!(
+        pre_walk,
+        walk
     );
-
-    walk::install();
-
-    attack::install();
-
-    attack_s3::install();
-
-    attack_s4::install();
-
-    attack_s4::install();
-
-    attack_air::install();
-    ladder_attack::install();
-    chargeshot::install();
-
-    special_n::install();
-    rockbuster::install();
-
-    special_s::install();
-
-    special_lw::install();
 }
