--- fighters/rockman/src/vl.rs
@@ -1,6 +1,0 @@
-pub mod private {
-    pub const CHARGE_SHOT_CLEAR_INPUT_FRAME : i32 = 6;
-    pub const CHARGE_SHOT_DELAY_CHARGE_FRAME : i32 = 50;
-    pub const CHARGE_SHOT_MAX_FRAME : i32 = 160;
-    pub const CHARGE_SHOT_RELEASE_FRAME : i32 = 6;
-}
